+++ README.md
@@ -1,29 +1,46 @@
-# ScrapingFinancialMarket
+# 📊 Financial Market Analyzer
 Recoge las empresas a mejor precio según el mercado de valores mediante scraping web
 
 Este proyecto solo sigue fines educativos y surge de la curiosidad de generar una herramienta que combine el mundo de las finanzas con el scraping web. En ningun caso es una recomendación de inversión. De igual modo, los datos son fiables y recogidos en tiempo real cuando se ejecute el script que contiene el core del proyecto.
 
-## Métricas a considerar
-* Indicador |  Descripción  |  Interpretación
-* Market Cap |  Valor de capitalización busátil | Mide el tamaño de la empresa
-* Enterprise Value  |  Valor total de la empresa (Incluyendo deuda)  |  La deuda se calcula con la diferencia de EV con respecto al market cap
-* Profit Margin  |  Margen de beneficios generados  |  Mide la solvencia de la empresa en términos monetarios
-* EBITDA  |  Beneficio bruto operativo  |  A mayor EBITDA, mayor potencial en la acción
 
-## Entorno técnico empleado para este proyecto
-* Python 3.11
-* Librerías scraping, financias y cálculos mat.: requests, BeatifulSoup, yfinance, pandas
+## 1. Introducción
 
-## Estructura del proyecto
-FinancialMarketsAnalysis/
-├── README.md
-├── requirements.txt
-├── main.py
-├── scraper/
-│   └── yahoo_scraper.py
-├── analysis/
-│   └── analyzer.py
-└── data/
-    └── tickers_sp500.csv
-    └── tickers_euroStock50.csv
-    └── tickers_ICEAsia.csv
+**Financial Market Analyzer** es una herramienta desarrollada en Python que permite analizar acciones de mercados bursátiles, especialmente del S&P500, utilizando datos financieros en tiempo real extraídos de Yahoo Finance. El sistema evalúa cada acción mediante distintos índices personalizados con el objetivo de detectar posibles oportunidades de inversión basadas en criterios fundamentales y de valor.
+
+---
+
+## 2. ¿De dónde surge la idea?
+
+La idea este proyecto nace de la curiosidad por entender mejor cómo se comportan los mercados financieros desde un enfoque cuantitativo. Me interesaba aplicar mis conocimientos técnicos, como el manejo de datos, automatización y diseño de sistemas, a un dominio real y complejo como es la bolsa de valores.
+La idea fue crear una herramienta práctica que me permitiera explorar datos financieros de forma automatizada, desarrollar métricas personalizadas.
+
+---
+
+## 3. ¿A quién va dirigido?
+
+Este proyecto está pensado para:
+
+- 🧑‍💻 Estudiantes de finanzas o ciencia de datos interesados en mercados bursátiles.
+- 📈 Inversores particulares que deseen analizar valores del EuroStoxx 50 con criterios cuantitativos.
+- 🧠 Desarrolladores que quieran aprender a combinar **Python** y **análisis financiero** en un proyecto realista y funcional.
+
+---
+
+## 4. ¿Cómo ejecutarlo?
+
+### ✅ Requisitos
+
+- Python 3.12.X
+- Acceso a internet (para consultas a Yahoo Finance)
+- Instalar las dependencias necesarias, a través del siguiente comando.
+
+### Instalación del código
+
+> git clone <>
+
+### 📦 Instalación de dependencias
+
+En la raíz del proyecto, ejecuta:
+
+> pip install -r requirements.txt
