+++ docs/index.md
@@ -1,0 +1,6 @@
+# Documentación Técnica del Proyecto
+
+Bienvenido a la documentación técnica del proyecto *Financial Market Analyzer*. Aquí encontrarás información detallada sobre su estructura interna, decisiones arquitectónicas y fundamentos de diseño.
+
+- [Arquitectura del Proyecto](architecture.md)
+- [Decisiones de Diseño](decisions.md)
