+++ requirements.txt
@@ -1,4 +1,4 @@
-yfinance==0.1.70
-pandas==1.3.3
-requests==2.26.0
-pytest==6.2.5
+yfinance==0.2.40
+pandas==2.2.2
+requests==2.31.0
+pytest==8.2.2
